requirements.txt
# Development
black==22.8.0
flake8==5.0.4
isort==5.10.1
pre-commit==2.17.0

# Deployment
requests==2.27.1
flask==2.0.3
flask-cors==3.0.10
pycryptodome==3.17
click==8.0.4
rich==12.6.0
rich-click==1.2.1
loguru==0.6.0
gunicorn==20.1.0
waitress==2.0.0
pydantic==1.9.2
